improve error handling for missing user-id header; extract types from entities route
